refs #5 Made the code a bit nicer and changed a directory name.
