+++ example_scripts/example_runScripts.py
@@ -76,7 +76,7 @@ use_errors = True
 for run_data in ["neutron", "nist"]:
 
     if run_data == "neutron":
-        group_suffix_names = ['neutron_data']
+        group_suffix_names = ['neutron']
         group_names = ["Neutron data"]
         results_per_group, results_dir = \
         fitBenchmarking(algorithm=algorithm, data_dir=neutron_data_dir,
