+++ fitbenchmarking/resproc/color_definitions.txt
