+++ fitbenchmarking/resproc/fitdetails_tbls.py
