+++ fitbenchmarking/resproc/misc.py
@@ -23,8 +23,11 @@ fitting results.
 # <https://github.com/mantidproject/fitbenchmarking>.
 # Code Documentation is available at: <http://doxygen.mantidproject.org>
 
+
 from __future__ import (absolute_import, division, print_function)
 import os
+
+from resproc import visual_pages
 
 
 def display_name_for_minimizers(names):
@@ -94,6 +97,38 @@ def build_items_links(comparison_type, comp_dim, using_errors):
         items_link = ''
 
     return items_link
+
+
+def create_linked_probs(results_per_test, group_name, results_dir):
+    """
+    Creates the problem names with links to the visual display pages
+    in rst.
+
+    @param results_per_test :: results object
+    @param group_name :: name of the problem group
+    @param results_dir :: directory in which the results are saved
+
+    @returns :: array of the problem names with the links in rst
+    """
+
+    # Count keeps track if it is the same problem but different starting point
+    prev_name = ''
+    count = 1
+
+    linked_problems = []
+    for test_idx, prob_results in enumerate(results_per_test):
+        name = results_per_test[test_idx][0].problem.name
+        if name == prev_name:
+            count += 1
+        else:
+            count = 1
+        prev_name = name
+        name_index = name + ' ' + str(count)
+        name = '`' + name_index + ' ' + \
+               visual_pages.create(prob_results, group_name, results_dir, count)
+        linked_problems.append(name)
+
+    return linked_problems
 
 
 def make_restables_dir(results_dir, group_name):
